hack out parse functionality
